Added Signup capabilities
Users are now able to sign up (Thanks for the help Bao).
TODO
Remove enums from models (they are not necessary when the multiplicity is established in the database itself
Populate the Ticket_Status and Ticket_Type tables with corresponding values following the User_Roles table.
Organize SQL scripts into their respective types so it's easier to run each